requirements.txt
# === LangChain Core Components ===
langchain-core
langchain
langchain-experimental

llama-index
# === LangChain Integrations ===
langchain-google-genai
langchain-community
langchain-qdrant
langchainhub
langchain-google-genai
langchain-groq
langchain-text-splitters
# === Data Loading and Processing ===
pypdf
pymupdf
unstructured
# === Vector Database ===
qdrant-client
faiss-cpu
numpy
# === Google SDK ===
google-genai
# Embeddings
sentence_transformers
hf_xet
transformers
torch
charset-normalizer
chardet

fastembed

# === RAGAS ===
ragas

langchain-ollama
